% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/betydb.R
\name{betydb}
\alias{betydb}
\alias{betydb_citations}
\alias{betydb_sites}
\alias{betydb_species}
\alias{betydb_traits}
\alias{betydb_traits_and_yields}
\alias{betydb_variables}
\alias{betydb_yields}
\title{Search for traits from BettyDB}
\usage{
betydb_traits(id = NULL, genus = NULL, species = NULL, fmt = "json",
  key = NULL, user = NULL, pwd = NULL, include = "variables", ...)

betydb_species(id = NULL, genus = NULL, species = NULL, fmt = "json",
  key = NULL, user = NULL, pwd = NULL, ...)

betydb_yields(id = NULL, genus = NULL, species = NULL, fmt = "json",
  key = NULL, user = NULL, pwd = NULL, ...)

betydb_citations(id = NULL, genus = NULL, species = NULL, fmt = "json",
  key = NULL, user = NULL, pwd = NULL, ...)

betydb_variables(id = NULL, genus = NULL, species = NULL, fmt = "json",
  key = NULL, user = NULL, pwd = NULL, ...)

betydb_sites(id = NULL, city = NULL, fmt = "json", key = NULL,
  user = NULL, pwd = NULL, ...)

betydb_traits_and_yields(id = NULL, sitename = NULL, genus = NULL,
  author = NULL, fmt = "json", key = NULL, user = NULL, pwd = NULL,
  ...)
}
\arguments{
\item{id}{(integer) One or more ids for a species, site, variable, etc.}

\item{genus}{(character) A genus name. Optional}

\item{species}{(character) A specific epithet. Optional}

\item{fmt}{(character) Format to return data in, one of json, xml, csv. Only json
currently supported.}

\item{key}{(character) An API key. Use this or user/pwd combo. Save in your
\code{.Rprofile} file as \code{betydb_key}. Optional}

\item{user,pwd}{(character) A user name and password. Use a user/pwd combo or an API key.
Save in your \code{.Rprofile} file as \code{betydb_user} and \code{betydb_pwd}. Optional}

\item{...}{Curl options passed on to \code{\link[httr]{GET}}. Optional}

\item{city}{(character) City name. Optional}
}
\description{
Search for traits from BettyDB
}
\details{
Details:
}
\section{Authentication}{

Defers to use API key first since it's simpler, but if you don't have
an API key, you can supply a username and password.
}

\section{Functions}{

Singular functions like \code{betydb_trait} accept an id and additional parameters,
and return a list of variable outputs depending on the inputs.

However, plural functions like \code{betydb_traits} accept query parameters, but not
ids, and always return a single data.frame.
}
\examples{
\dontrun{
# Search with parameters
## Traits
(out <- betydb_traits(genus = 'Miscanthus', species = "giganteus"))
library("dplyr")
out \%>\%
 group_by(specie_id) \%>\%
 summarise(mean_result = mean(as.numeric(mean), na.rm = TRUE)) \%>\%
 arrange(desc(mean_result))

## Species
betydb_species(genus = 'Miscanthus')

## Yields
betydb_yields(genus = 'Miscanthus')

## Citations
betydb_citations(genus = 'Miscanthus')

## Sites
betydb_sites(city = 'Beddington')

# Get by ID
## Traits
betydb_traits(id = 10) or betydb_traits(10)
## Species
betydb_species(id = 1)
## Citations
betydb_citations(id = 1)
## Citations
betydb_sites(id = 795)
}
}
\author{
Scott Chamberlain \email{myrmecocystus@gmail.com}
}
\references{
API documentation \url{https://www.authorea.com/users/5574/articles/7062}
}

