% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betydb.R
\name{betydb_query}
\alias{betydb_query}
\title{Query a BETY table}
\usage{
betydb_query(..., table = "search", key = NULL, api_version = "v0",
  betyurl = "https://www.betydb.org/")
}
\arguments{
\item{...}{(named character) Columns to query, as key="value" pairs. Note that betydb_query passes these along to BETY with no check whether the requested keys exist in the specified table.}

\item{table}{(character) The name of the database table to query, or "search" (the default) for the traits and yields view}

\item{key}{(character) An API key. Use this or user/pwd combo. Save in your
\code{.Rprofile} file as \code{betydb_key}. Optional}

\item{api_version}{(character) Which version of the betydb api to use? Optional, defaults to 'v0'}

\item{betyurl}{(string) url to target instance of betydb. Default is https:/www.betydb.org/}
}
\value{
A data.frame with attributes containing request metadata, or NULL if the query produced no results
}
\description{
Query a BETY table
}
\details{
Use betydb_query to retrieve records from a table that match on all the column filters specified in '...'.
If no filters are specified, retrieves the whole table. In API versions that support it (i.e. not in v0), filter strings beginning with "~" are treated as regular expressions.
}
\examples{
\dontrun{
# literal vs regular expression vs anchored regular expression:
betydb_query(units="Mg", table="variables")
# NULL
betydb_query(units="Mg/ha", table="variables") \%>\% select(name) \%>\% c()
# $name
# [1] "a_biomass"                  "root_live_biomass"
# [3] "leaf_dead_biomass_in_Mg_ha" "SDM"

betydb_query(genus="Miscanthus", table="species") \%>\% nrow()
# [1] 10
(betydb_query(genus="~misc", table="species", api_version="beta")
 \%>\% select(genus)
 \%>\% unique() \%>\% c())
# $genus
# [1] "Platymiscium" "Miscanthus"   "Dermiscellum"

(betydb_query(genus="~^misc", table="species", api_version="beta")
 \%>\% select(genus)
 \%>\% unique() \%>\% c())
# $genus
# [1] "Miscanthus"
}

}

