% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coral.R
\name{coral-defunct}
\alias{coral-defunct}
\alias{coral_taxa}
\alias{coral_traits}
\alias{coral_locations}
\alias{coral_methodologies}
\alias{coral_resources}
\alias{coral_species}
\title{Search for coral data on coraltraits.org}
\usage{
coral_taxa(...)

coral_traits(...)

coral_locations(...)

coral_methodologies(...)

coral_resources(...)

coral_species(...)
}
\arguments{
\item{...}{ignored}
}
\description{
DEFUNCT: As far as we can tell the coraltraits.org API is down
for good
}
\references{
https://coraltraits.org/
}
\keyword{internal}
