% Generated by roxygen2 (4.0.2): do not edit by hand
\name{coral}
\alias{coral}
\alias{coral_locations}
\alias{coral_methodologies}
\alias{coral_resources}
\alias{coral_taxa}
\alias{coral_traits}
\title{Search for coral data on http://coraltraits.org/.}
\usage{
coral_taxa(taxon, taxonomy = "off", contextual = "on", global = "off",
  ...)

coral_traits(trait, taxonomy = "off", contextual = "on", global = "off",
  ...)

coral_locations(location, taxonomy = "off", contextual = "on",
  global = "off", ...)

coral_methodologies(methodology, taxonomy = "off", contextual = "on",
  global = "off", ...)

coral_resources(resource, taxonomy = "off", contextual = "on",
  global = "off", ...)
}
\arguments{
\item{taxonomy}{xxx}

\item{contextual}{xxx}

\item{global}{xxx}

\item{...}{Curl options passed on to \code{\link[httr]{GET}}}

\item{trait}{xxx}

\item{coral}{xxx}

\item{coral}{xxx}
}
\description{
Search for coral data on http://coraltraits.org/.
}
\examples{
\donttest{
# Get data by taxon
df <- coral_taxa(80)
head(df)

# Get data by trait
df <- coral_traits(105)
head(df)

# Get data by methodology
df <- coral_methodologies(2)
head(df)

# Get data by location
df <- coral_locations(132)
head(df)

# Get data by resource
df <- coral_resources(10)
head(df)
}
}
\references{
See info for each data source at \url{http://coraltraits.org/}
}

