% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g_invasive.R
\name{g_invasive}
\alias{g_invasive}
\title{Check invasive species status for a set of species from GISD database}
\usage{
g_invasive(...)
}
\description{
Check invasive species status for a set of species from GISD database
}
\keyword{internal}
