% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/birdlife.R
\name{get_birdlife_habitat}
\alias{get_birdlife_habitat}
\title{Get bird habitat information from BirdLife/IUCN}
\usage{
get_birdlife_habitat(id)
}
\arguments{
\item{id}{IUCN species ID}
}
\value{
a \code{data.frame} with level 1 and level 2 habitat classes, as well as importance
ratings and occurrence type (e.g. breeding or non-breeding).  The habitat classification
scheme is described at http://www.iucnredlist.org/technical-documents/classification-schemes/habitats-classification-scheme-ver3
}
\description{
Get bird habitat information from BirdLife/IUCN
}

