% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/leda.R
\name{leda}
\alias{leda}
\title{Access LEDA trait data}
\usage{
leda(trait = "age_first_flowering", ...)
}
\arguments{
\item{trait}{(character) Trait to get. See Details.}

\item{...}{Curl options passed on to \code{\link[httr]{GET}}}
}
\description{
Access LEDA trait data
}
\details{
For parameter \code{trait}, one of age_first_flowering, branching,
buds_seasonality, buds_vertical_dist, buoyancy, canopy_height, dispersal_type,
leaf_distribution, ldmc_geo, leaf_mass, leaf_size, morphology_disperal,
growth_form, life_span, releasing_height, seed_bank, seed_longevity,
seed_mass, seed_number, seed_shape, shoot_growth_form, sla_geo, snp, ssd, or tv
}
\examples{
\dontrun{
# Age of first flowering
leda(trait = "age_first_flowering")

# Age of first flowering
leda(trait = "seed_bank")
}
}
\author{
Scott Chamberlain \email{myrmecocystus@gmail.com}
}

