% Generated by roxygen2 (4.0.2): do not edit by hand
\name{traitbank}
\alias{traitbank}
\title{Search for traits from EOL's Traitbank.}
\usage{
traitbank(trait, cache_ttl = NULL, ...)
}
\arguments{
\item{trait}{A trait id}

\item{cache_ttl}{Cache code}

\item{...}{Curl options passed on to \code{\link[httr]{GET}}}
}
\description{
Search for traits from EOL's Traitbank.
}
\examples{
\donttest{
traitbank(trait = 328574)
}
}
\author{
Scott Chamberlain \email{myrmecocystus@gmail.com}
}
\references{
\url{http://eol.org/info/516}
}

