% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defunct.R
\name{traits-defunct}
\alias{traits-defunct}
\title{Defunct functions in traits}
\description{
These functions have been removed. See package \code{originr}
}
\details{
\itemize{
\item \code{eol_invasive_}: This function has moved to a new package.
See \code{originr::eol}
\item \code{fe_native}: This function has moved to a new package.
See \code{originr::flora_europaea}
\item \code{g_invasive}: This function has moved to a new package.
See \code{originr::gisd}
\item \code{is_native}: This function has moved to a new package. See
\code{originr::is_native}
\item \code{tr_usda}: the API behind this function is down for good
}
}
