% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{traits-deprecated}
\alias{traits-deprecated}
\title{Deprecated functions in traits}
\description{
These functions still work but will be removed (defunct) in the next version.
}
\details{
\itemize{
 \item \code{\link{eol_invasive_}}: This function is moving to a new package to
 be on CRAN soon. It will be \code{eol} in \code{originr} package
 \item \code{\link{fe_native}}: This function is moving to a new package to
 be on CRAN soon. It will be \code{flora_europaea} in \code{originr} package
 \item \code{\link{g_invasive}}: This function is moving to a new package to
 be on CRAN soon. It will be \code{gisd} in \code{originr} package
 \item \code{\link{is_native}}: This function is moving to a new package to
 be on CRAN soon. It will be \code{is_native} in \code{originr} package
}
}

