% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traits-package.r
\docType{package}
\name{traits-package}
\alias{traits-package}
\alias{traits}
\title{traits - Species trait data from around the web}
\description{
Currently included in \code{traits} with the associated function name or
function prefix:
}
\details{
\itemize{
\item BETYdb http://www.betydb.org - \code{betydb_}
\item National Center for Biotechnology Information - NCBI
http://www.ncbi.nlm.nih.gov/ - \code{ncbi_}
\item Encyclopedia of Life Traitbank - \code{traitbank_}
\item Birdlife International https://www.birdlife.org/ -
\code{birdlife_}
\item LEDA Traitbase http://www.leda-traitbase.org/LEDAportal/index.jsp -
\code{leda_}
\item Zanne et al. plant dataset - \code{\link[=tr_zanne]{tr_zanne()}}
\item Amniote life history dataset - \code{\link[=tr_ernest]{tr_ernest()}}
}

See also \link{traits-defunct}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://docs.ropensci.org/traits/}
  \item \url{https://github.com/ropensci/traits}
  \item Report bugs at \url{https://github.com/ropensci/traits/issues}
}

}
\author{
\strong{Maintainer}: David LeBauer \email{dlebauer@gmail.com} (\href{https://orcid.org/0000-0001-7228-053X}{ORCID})

Authors:
\itemize{
  \item Scott Chamberlain \email{myrmecocystus@gmail.com} (\href{https://orcid.org/0000-0003-1444-9135}{ORCID}) [copyright holder]
  \item Zachary Foster
  \item Ignasi Bartomeus
  \item Chris Black
  \item David Harris
}

Other contributors:
\itemize{
  \item Rupert Collins [contributor]
}

}
\keyword{internal}
