% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/betydb.R
\name{betydb}
\alias{bety_traits}
\alias{betydb}
\title{Search for traits from BettyDB}
\usage{
bety_traits(genus = NULL, species = NULL, fmt = "json", key = NULL,
  user = NULL, pwd = NULL, ...)
}
\arguments{
\item{genus}{(character) A genus name. Optional}

\item{species}{(character) A specific epithet. Optional}

\item{fmt}{(character) Format to return data in, one of json, xml, csv. Only json
currently supported.}

\item{key}{(character) An API key. Use this or user/pwd combo. Optional}

\item{user,pwd}{(character) A user name and password. Use a user/pwd combo or an API key.
Optional}

\item{...}{Curl options passed on to \code{\link[httr]{GET}}. Optional}
}
\description{
Search for traits from BettyDB
}
\details{
Authentication defers to use API key first since it's simpler, but if you don't have
an API key, you can supply a username and password.
}
\examples{
\dontrun{
out <- bety_traits(genus = 'Miscanthus', species = "giganteus")
head(out)
}
}
\author{
Scott Chamberlain \email{myrmecocystus@gmail.com}
}
\references{
API documentation \url{https://www.authorea.com/users/5574/articles/7062}
}

