% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/traitbank.R
\name{traitbank}
\alias{traitbank}
\title{Search for traits from EOL's Traitbank.}
\usage{
traitbank(trait, cache_ttl = NULL, ...)
}
\arguments{
\item{trait}{A page id. I know, not ideal. Would be better if this was a trait
id or trait name.}

\item{cache_ttl}{Cache code}

\item{...}{Curl options passed on to \code{\link[httr]{GET}}}
}
\description{
Search for traits from EOL's Traitbank.
}
\details{
See \url{http://eol.org/data_glossary} for human readable definitions for
the attribute terms that EOL uses. Go to \url{http://eol.org/data_search} for the
web interface to Traitbank.
}
\examples{
\donttest{
res <- traitbank(trait = 328574)
res$context
names( res$graph )
head( res$graph )
}
}
\author{
Scott Chamberlain \email{myrmecocystus@gmail.com}
}
\references{
\url{http://eol.org/info/516}
}

