% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defunct.R
\name{traits-defunct}
\alias{traits-defunct}
\title{Defunct functions in traits}
\description{
These functions have been removed. See package \code{originr}.
}
\details{
\itemize{
 \item \code{\link{eol_invasive_}}: This function has moved to a new package.
 See \code{eol} in \code{originr}
 \item \code{\link{fe_native}}: This function has moved to a new package.
 See \code{flora_europaea} in \code{originr}
 \item \code{\link{g_invasive}}: This function has moved to a new package.
 See \code{gisd} in \code{originr}
 \item \code{\link{is_native}}: This function has moved to a new package.
 See \code{is_native} in \code{originr}
}
}
