% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/traits-package.r
\docType{package}
\name{traits-package}
\alias{traits}
\alias{traits-package}
\title{Species trait data from around the web}
\description{
traits
}
\author{
Ignasi Bartomeus \email{nacho.bartomeus@gmail.com}

Zachary Foster \email{zacharyfoster1989@gmail.com}

David LeBauer \email{dlebauer@gmail.com}

Scott Chamberlain \email{myrmecocystus@gmail.com}
}
\keyword{package}

