% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traits-package.r
\docType{package}
\name{traits-package}
\alias{traits}
\alias{traits-package}
\title{traits - Species trait data from around the web}
\description{
Currrently included in \code{traits} with the associated function prefix:
\itemize{
 \item BETYdb \url{http://www.betydb.org} - \code{betydb_}
 \item National Center for Biotechnology Information - NCBI
 \url{http://www.ncbi.nlm.nih.gov/} - \code{ncbi_}
 \item Global Invasive Species Database - GISD \url{http://www.issg.org/database/welcome/}
 - \code{g_}
 \item Encyclopedia of Life Invasive Species - \code{eol_}
 \item Encyclopedia of Life Traitbank - \code{traitbank_}
 \item Coral Traits Database \url{http://coraltraits.org/} - \code{coral_}
 \item Flora Europaea \url{http://rbg-web2.rbge.org.uk/FE/fe.html} - \code{fe_}
 \item Birdlife International \url{http://rbg-web2.rbge.org.uk/FE/fe.html} - \code{birdlife_}
 \item LEDA Traitbase http://www.leda-traitbase.org/LEDAportal/index.jsp - \code{leda_}
 \item More to come ...
}
}
\examples{
\dontrun{
library("traits")

## Search the Coral database
## Get the species list and their ids
coral_species()

## Get data by taxon
coral_taxa(80)
}

}
\author{
Ignasi Bartomeus \email{nacho.bartomeus@gmail.com}

Zachary Foster \email{zacharyfoster1989@gmail.com}

David LeBauer \email{dlebauer@gmail.com}

David Harris \email{davharris@ucdavis.edu}

Scott Chamberlain \email{myrmecocystus@gmail.com}
}
\keyword{package}

