% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traits-package.r
\docType{package}
\name{traits-package}
\alias{traits-package}
\alias{traits}
\title{traits - Species trait data from around the web}
\description{
Currently included in \code{traits} with the associated function name or
function prefix:
}
\details{
\itemize{
\item BETYdb http://www.betydb.org - \code{betydb_}
\item National Center for Biotechnology Information - NCBI
http://www.ncbi.nlm.nih.gov/ - \code{ncbi_}
\item Encyclopedia of Life Traitbank - \code{traitbank_}
\item Birdlife International https://www.birdlife.org/ -
\code{birdlife_}
\item LEDA Traitbase http://www.leda-traitbase.org/LEDAportal/index.jsp -
\code{leda_}
\item Zanne et al. plant dataset - \code{\link[=tr_zanne]{tr_zanne()}}
\item Amniote life history dataset - \code{\link[=tr_ernest]{tr_ernest()}}
}

See also \link{traits-defunct}
}
\author{
Scott Chamberlain

Ignasi Bartomeus

Zachary Foster

David LeBauer

David Harris

Rupert Collins
}
\keyword{package}
